🎨  Upgraded some styling
Still need a lot of work before we're production ready
