+++ src/pages/dashboard/DashboardRoot.tsx
@@ -3,6 +3,9 @@ import { useQuery } from "@tanstack/react-query";
 import { getAdminDataQuery } from "../../utils/firebaseQueries.ts";
 import { DashboardHeader } from "./DashboardHeader.tsx";
 
+import styles from "./sass/Dashboard.module.scss";
+import EmployeeDashboard from "./EmployeeDashboard.tsx";
+
 export function DashboardRoot() {
   const currentUserId = auth?.currentUser?.uid ?? "";
 
@@ -10,18 +13,23 @@ export function DashboardRoot() {
   // TODO: test for no admin data. how to handle? consider going to login
 
   return (
-    <div className="dashboard-container">
+    <div className={styles.dashboardContainer}>
       <DashboardHeader />
       {!adminData ? (
         <></>
       ) : adminData.omniAdmin ? (
-        <div>omniAdmin</div>
+        <div>
+          omniAdmin
+          {currentUserId}
+        </div>
       ) : adminData.isAdmin ? (
-        <div>is admin</div>
+        <div>
+          is admin
+          {currentUserId}
+        </div>
       ) : (
-        <div>is not admin</div>
+        <EmployeeDashboard />
       )}
-      {currentUserId}
     </div>
   );
 }
