+++ src/pages/dashboard/sass/Dashboard.module.scss
@@ -1,0 +1,16 @@
+@use "../../../variables.scss" as *;
+
+.dashboardContainer {
+  @include flex-default(column);
+
+  min-height: 100vh;
+  font-family: Arial, sans-serif;
+  background-color: $bg-secondary;
+  position: relative;
+
+  /* this serves to push the hour adder down some, since it was actually weird for it to be dead center */
+  /* also, that's how tall the header is (below) */
+  padding-top: 75px;
+  padding-bottom: 30px;
+  box-sizing: border-box; /* this prevents that ^ from forcing the window to be scroll-able */
+}
